refactor(tests): Restructure dirs and add guide
