--- testing/l2/string.py
@@ -1,3 +1,0 @@
-def str_to_felt(text):
-    b_text = bytes(text, 'UTF-8')
-    return int.from_bytes(b_text, "big")
