+++ tests/account/account_contract_test.py
@@ -3,10 +3,13 @@ import asyncio
 from starkware.starknet.testing.starknet import Starknet
 from starkware.starkware_utils.error_handling import StarkException
 from starkware.starknet.definitions.error_codes import StarknetErrorCode
-from utils.Signer import Signer
+from tests.utils import Signer
 
 signer = Signer(123456789987654321)
 other = Signer(987654321123456789)
+
+CAIRO_ACCOUNT_CONTRACT = "contracts/openzeppelin/account/Account.cairo"
+CAIRO_INITIALIZABLE_CONTRACT = "contracts/Initializable.cairo"
 
 
 @pytest.fixture(scope='module')
@@ -18,7 +21,7 @@ def event_loop():
 async def account_factory():
     starknet = await Starknet.empty()
     account = await starknet.deploy(
-        "contracts/l2/utils/Account.cairo",
+        CAIRO_ACCOUNT_CONTRACT,
         constructor_calldata=[signer.public_key]
     )
 
@@ -36,7 +39,7 @@ async def test_constructor(account_factory):
 @pytest.mark.asyncio
 async def test_execute(account_factory):
     starknet, account = account_factory
-    initializable = await starknet.deploy("contracts/l2/utils/Initializable.cairo")
+    initializable = await starknet.deploy(CAIRO_INITIALIZABLE_CONTRACT)
 
     execution_info = await initializable.initialized().call()
     assert execution_info.result == (0,)
@@ -50,7 +53,7 @@ async def test_execute(account_factory):
 @pytest.mark.asyncio
 async def test_return_value(account_factory):
     starknet, account = account_factory
-    initializable = await starknet.deploy("contracts/l2/utils/Initializable.cairo")
+    initializable = await starknet.deploy(CAIRO_INITIALIZABLE_CONTRACT)
 
     # initialize, set `initialized = 1`
     await signer.send_transaction(account, initializable.contract_address, 'initialize', [])
@@ -64,7 +67,7 @@ async def test_return_value(account_factory):
 @ pytest.mark.asyncio
 async def test_nonce(account_factory):
     starknet, account = account_factory
-    initializable = await starknet.deploy("contracts/l2/utils/Initializable.cairo")
+    initializable = await starknet.deploy(CAIRO_INITIALIZABLE_CONTRACT)
     execution_info = await account.get_nonce().call()
     current_nonce = execution_info.result.res
