+++ tests/conftest.py
@@ -2,7 +2,7 @@ import asyncio
 import pytest
 import dill
 import os
-from utils import Signer, uint, str_to_felt, MAX_UINT256, assert_revert
+from tests.utils import Signer, uint, str_to_felt, MAX_UINT256, assert_revert
 import sys
 import time
 
@@ -15,6 +15,8 @@ from starkware.starknet.business_logic.state import BlockInfo
 
 sys.stdout = sys.stderr
 
+# Create signers that use a private key to sign transaction objects.
+DUMMY_PRIVATE = 123456789987654321
 
 CONTRACT_SRC = os.path.join(os.path.dirname(__file__), "../..", "contracts")
 INITIAL_LORDS_SUPPLY = 500000000 * (10 ** 18)
@@ -280,7 +282,7 @@ async def l06_combat(starknet, xoroshiro) -> StarknetContract:
 
 @pytest.fixture(scope="module")
 async def l06_combat_tests(starknet, xoroshiro) -> StarknetContract:
-    contract = compile("testing/l2/L06_Combat_tests.cairo")
+    contract = compile("tests/settling_game/L06_Combat_tests.cairo")
     # a quirk of the testing framework, even though the L06_Combat_tests contract
     # doesn't have a constructor, it somehow calls (I guess) the constructor of
     # L06_Combat because it imports from it; hence when calling deploy, we need
@@ -296,5 +298,40 @@ async def s06_combat(starknet) -> StarknetContract:
 
 @pytest.fixture(scope="module")
 async def s06_combat_tests(starknet) -> StarknetContract:
-    contract = compile("testing/l2/S06_Combat_tests.cairo")
+    contract = compile("tests/settling_game/S06_Combat_tests.cairo")
     return await starknet.deploy(contract_def=contract)
+
+@pytest.fixture(scope='module')
+async def account_factory(request):
+    num_signers = request.param.get("num_signers", "1")
+    starknet = await Starknet.empty()
+    accounts = []
+    signers = []
+
+    print(f'Deploying {num_signers} accounts...')
+    for i in range(num_signers):
+        signer = Signer(DUMMY_PRIVATE + i)
+        signers.append(signer)
+        account = await starknet.deploy(
+            "contracts/openzeppelin/account/Account.cairo",
+            constructor_calldata=[signer.public_key]
+        )
+        accounts.append(account)
+
+        print(f'Account {i} is: {hex(account.contract_address)}')
+
+    # Initialize network
+
+    # Admin is usually accounts[0], user_1 = accounts[1].
+    # To build a transaction to call func_xyz(arg_1, arg_2)
+    # on a TargetContract:
+
+    # await Signer.send_transaction(
+    #   account=accounts[1],
+    #   to=TargetContract,
+    #   selector_name='func_xyz',
+    #   calldata=[arg_1, arg_2],
+    #   nonce=current_nonce)
+
+    # Note that nonce is an optional argument.
+    return starknet, accounts, signers
