+++ tests/nft_marketplace/nft_marketplace_test.py
@@ -1,9 +1,6 @@
 import pytest
-import asyncio
-from utils import Signer, uint, str_to_felt, MAX_UINT256, assert_revert
-from starkware.starknet.testing.starknet import Starknet
-from starkware.starkware_utils.error_handling import StarkException
-from starkware.starknet.definitions.error_codes import StarknetErrorCode
+from tests.utils import str_to_felt, uint
+
 
 # Params
 BASE_URI = str_to_felt('https://realms.digital/token/')
