+++ tests/settling_game/06_combat_test.py
