+++ tests/settling_game/L06_Combat_tests.cairo
