+++ tests/settling_game/S06_Combat_tests.cairo
